[unnecessary copy] Only track valid returned addresses for modifications
Summary:
In D38908357, we have started keeping track of returned value-> actuals for unknown calls so that if the returned value is modified, we can check for modifications to actuals. However, if the actual has gone out of scope (i.e. invalidated), we shouldn't keep track of it unnecessarily.

I can't come up with a simple example in this diff but this became an issue in the next diff where we update the ranking for timestamps. Without this diff, we end up not reporting on the intermediate copy for the following example

```
extern std::string __infer_taint_source();
extern void __infer_taint_sink(std::string);

void replace1_bad() {
  auto source = __infer_taint_source();
  __infer_taint_sink(std::string("string").replace(0, 5, source));
}

```

because when we are checking for modifications to the result of `std::string("string").replace(0, 5, source)`, we lookup the actuals originating from replace and find that the first argument `std::string("string")` is getting modified but it eventually goes out of scope.

Reviewed By: skcho

Differential Revision: D46686288

fbshipit-source-id: cc3f72e487a43db726f26967cd3f42ce6700d2c8