[unnecessary copy] Distinguish WrittenTo by timestamp
Summary:
`WrittenTo` attribute keeps track of timestamp but this timestamp is ignored when we are are adding new attributes (due to Ranking in the Attribute set which just does a simple add instead of update since ranks match). This causes FPs where we miss to recognize that a value is modified after it is copied because we ignored the subsequent modifications tracked by the timestamp.

This diff fixes the issue by doing a proper update that replaces the old WrittenTo.

Reviewed By: dulmarod

Differential Revision: D46676699

fbshipit-source-id: e976be86bdee4c564753fb5470f5ba56a7fb6954