[sqlite] add immutable flag
Summary:
The `extract-capture-from` and `complete-capture-from` functionality read from an input capture DB which is not modified. That DB is currently attached as normal, which means read locks are set in place when querying. At high degrees of parallelism, this results in IO errors.

Adding the `immutable` flag to `ATTACH` tells Sqlite to (a) treat this as a read-only DB and (b) assume no other process will change it. This allows increased parallelism.

NB the input DB must have had a `wal_checkpoint` performed before attaching, as anything in the WAL files that's not committed to the main DB will be ignored.

Reviewed By: jvillard

Differential Revision: D46255408

fbshipit-source-id: 8f3250993face8ba8bc4d80bca46f3a2698c7723