[pulse] Check Recoverable result for ContinueProgram in pulse-force-continue
Summary:
The intention of `--pulse-force-continue` was to treat the function as "unknown" when there are no `ContinueProgram` disjuncts. However, the code was checking only the `Ok` case of `pulse_result` and didn't check `Recoverable` which can also contain `ContinueProgram` disjuncts.

E.g. a taint error when applying a callee summary would produce a `Recoverable` result but otherwise the state would have a bunch of `ContinueProgram` disjuncts. There's no reason to treat the callee as unknown in this case.

Reviewed By: dulmarod

Differential Revision: D46806044

fbshipit-source-id: 6a48561b7da4dd69e2e943531e2c00be5d32622a