[taint] Add a repro for a taint FP due to abstract value unification
Summary:
The setup of this repro is the following:
1. we have 2 booleans: one tainted and one is not,
2. the control flow is such that the two booleans are getting 0 val and
   hence unified.
3. the untainted boolean flows into a taint sink.

Because of p2 the attributes of the two bools get merged and the
untainted bool will be considered as tainted during taint check.

As discussed with jvillard we cannot not unify the bools. Instead we should
handle taint checks similarly to what we do with nullptr dereferences:
1. taint events should be introduced and added to the value history
2. when we check for taint we use attributes to find which values are
   *potentially* tainted and then use ValueHistory to check if they are
   truly tainted.

This issue will be addressed in follow up diffs.

Reviewed By: dulmarod

Differential Revision: D46842676

fbshipit-source-id: b2d30684f5e850ad5d3f59558ea25ab95589860e