[pulse] fix debug procedures summary printing
Summary:
The printing functionality in Pulse depends on knowing what the current procdesc is, but there is no such thing when trying to print pulse summaries from the `debug` command, making it crash.

This diff fixes that by exposing the global state storing the current procdesc as an option type.

Reviewed By: ezgicicek

Differential Revision:
D46889094

Privacy Context Container: L1122176

fbshipit-source-id: 05cbf710226bf6b51c81399fabd0268b95441085