[infer/python] first support for imports.
Summary:
This diff introduces a first step in the world of Python's `import`. We support the very basic `import foo`. No complex path, no `from`, no `as` for now or other subtlety of `importlib`

Imports are not easy, because the compiler doesn't now if a name is a module or an instance of a class. So it will compile both
```
import base
base.f()
```
and
```
class C:
   def f():
      pass

c = C()
c.f()
```

in a similar fashion: `base.f()` and `c.f()` will use a similar combinaison of `LOAD_METHOD` and `CALL_METHOD`. So we use dedicated cell types and symbols to keep track of as much information as we can to avoid too many dynamic behavior.

Reviewed By: davidpichardie

Differential Revision: D46794900

fbshipit-source-id: 4b77eeb6456abbdc22cc2e7d5ad0cf2204ff2dfc