[infer/python] support for user denotable class names
Summary:
Up until now, we didn't correctly supported class names as type. This
first diff improves the situation by looking into our symbol environment
to see if the source string refers to any type we might be aware of.

If the type is unknown, we currently output `PyObject`. We'll probably
revisit this decision once we run into real life examples with import,
and leverage the `Tenv` merging to find the right types.

Reviewed By: davidpichardie

Differential Revision: D47094718

fbshipit-source-id: 0486f395aaaf167ac1ec0f6b5c2e69839d4cb960