[clang] Improve the translation of NSArray.objectEnumerator calls
Summary:
The frontend translates the loops of the kind `for (NSArray* item in items)` via the method `NSArray.objectEnumerator` which is what happens internally. However, because we create those method calls in the AST ourselves, we don't have the correct pointer to find the correct definition of the method in the AST to get its signature from. So we end up in the default case where we don't know return type or arguments.

Because of this, because this method was an instance method called with `self` but with 0 formals, the biabduction in Pulse was failing.

Reviewed By: ngorogiannis

Differential Revision: D47126603

fbshipit-source-id: 536e84e2a8a7196224820c844f4e77264bdc0e90