[infer/hack][experimental] update Tenv.resolve_method
Summary:
With `experimental_self_parent_in_trait`, the arity of method in traits
is changed, by adding a new argument at then end: `self`

This diff updates `Tenv.resolve_method` to correctly patch the arity of
procnames while we are looking for a method in a parent. It also forward
more into to `pulse`, namely in which class a trait was used. This
enable us to create a class names to provide for `self` (the class
where the trait was used)

Reviewed By: davidpichardie

Differential Revision: D46513855

fbshipit-source-id: fed251129f8c9d23e99704c14c3bb303bad72b64