[pulse] Fix copy ctor model for C++'s shared pointers
Summary: The pulse models for  `shared_ptr`'s copy ctor  was calling the destructor helper function but this should only be done for the copy assignment on the object that is re-assigned (to decrease the reference count of the managed object). Copy ctor doesn't reduce any reference counts at all.

Reviewed By: skcho

Differential Revision: D47123883

fbshipit-source-id: 1b0574d93c02bfdff169e32e72db622bd9f717a6