[taint] Split field matchers from getters and setters
Summary:
This diff fixes an issue that we were not separating getters matchers from setters matchers, thus adding wrong taint attributes to values.

Now we split the field matchers in getters and setters, and when analysing `load` we use the getters matchers, and when analysing `store` we use the setters matchers.

Reviewed By: skcho

Differential Revision: D47190795

fbshipit-source-id: f102657c423b08f8d5fd22860f9fec9e60b43b74