[summary] do not crash when summary is too big for sqlite
Summary:
When the size of a value is greater than some threshold (2Gb I believe), Sqlite refuses to write it in the DB and throws an exception. Instead of crashing, we write the empty summary and return it so that it can be cached consistently.

To do this, `DBWriter` serialises the exception and sends it back to the client process (if the daemon is in use) so that it can be propagated up the client's stack and caught.

Reviewed By: jvillard

Differential Revision: D47190459

fbshipit-source-id: b72f3ee63563f224e280e3ba16eea7385c0071df