[pulse] Log more info when found a cycle during restoring formals
Summary:
Background: In pulse, it restores the parameter values that could be modified in a procedure when
making a summary. (D28001229) During the restoration, if it finds a cycle, it logs the problematic
abstract memory, because which means that something wrong/unexpected things happened.

This diff logs more information, procedure name and its location, for better debugging.

Reviewed By: ngorogiannis

Differential Revision: D47220945

fbshipit-source-id: d3242b73ce290192724717cefb2f1c09d1bd1914