[textual] adding a new if-then-else terminator
Summary:
When writing models in Textual, it is quite error-prone to deal with
complex boolean expressions. We had a new node terminator that allows high-level
boolean expressions. We add a parser and a lowering transformation that replaces
such an instruction with a jump instruction and extra nodes to compile the lazy
semantics of && and || operators.

The negation operator is used as an intermediate step during the translation, but
not directly provided to the programmer since she can just swap the then-else branches
instead.

Reviewed By: jvillard

Differential Revision: D47204513

fbshipit-source-id: da0d5f4829357068baf96c28ace5f4031df3df16