[taint] Start injecting taint events to value history
Summary:
With the necessary plumbing in place we can finally start injecting taint events to the value
history.

Ultimately, we'd like to rely on value history to decide when to report taint issues. For now,
however, the main chunks of logic still rely on attributes. The diff adds a simple history check,
though, gated by a new `--pulse-taint-check-history` flag. So, the analysis behavior shouldn't
change when the flag is not set, but we can enable flag in hack tests and start experimenting with
it.

Regarding checking history for taint: this is a temporary simplified check used to filter out
tainting cases where the taint attribute got attached to an untainted value due to this value being
unified with some other tainted value.

This needs to be improved in the following way:
1. More taint info needs to be shifted from attributes to value histories. In particular,
sanitization events should be introduced and added to the history.
2. This check should take into account the sequence of tainting and sanitization in the history
and also the kinds of taint that need to be checked against the sink.

Reviewed By: vsiles

Differential Revision: D47591785

fbshipit-source-id: aabeaa4632c43773135ad758ac93727e357e0070