[infer/python] store function parameter names and static status in the env
Summary:
To implement correctly some part of the class hierarchy related code,
I'll need to know what `self` is. In Python, `self` is the name of the
first parameter of a method. There is no special keyword for it.

This code is legal Python code
```
class C:
  def __init__(my_own_state):
     my_own_state.x = 42
```

We also store the fact that the current method is static or not: static
methods don't have `self`, so we need to be aware of that for our next
diff

Reviewed By: ezgicicek

Differential Revision: D47797612

fbshipit-source-id: a256266f96417e48fe87a459a32984cad842b149