[infer/python] generalize TwoWay jumps
Summary:
Until now, all TwoWay jumps had the same SSA arity.
While working on `JUMP_IF_TRUE_OR_POP` and al, this is no longer true:
the top of the stack is popped only in one branch.

This diffs generalizes the TwoWay constructor to account for different
arity. This is not used yet.

Reviewed By: davidpichardie

Differential Revision: D47984823

fbshipit-source-id: 621ce6b37cf3ffbd6b70c3a8ae8cbe736d8f9598