[pulse] Havoc std::__detail::Node_iterator typed interator
Summary:
Pulse is havoc'ing when an iterator is given to unknown function in
order to avoid false positives. For example,

* when an iterator is passed to its *next* function, such as `iter++`,
* and the next function is unknown,

unless havocing the iterator, Pulse will think all the elements got
from the iterator are the same, which will result in unexpected
abstract states and false positives in the loop.

This diff adds an iterator type to havoc:
`std::__detail::_Node_iterator`.

Reviewed By: jvillard

Differential Revision: D47996976

fbshipit-source-id: d4a24b2b4e9c88782628ac5aeef2a00c0f06ae57