[infer/python] constructor improved (support implicit ones)
Summary:
In Python, one can omit to write an explicit class constructor. If the class has a parent, the implicit behavior will be to call the first constructor available in the parent class hierarchy:
```
class A:
  def __init__(self):
    print("A")

class B(A):
  pass

class C(B):
 def __init__(self):
   super().__init__()  # [0]
   print("C")

c = C() # will print "A" and "C"
```

However if these declarations are in different files, it is not possible for us to generate valid stubs that does the allocation and call in the right constructor, because we won't know which parent class constructor to call, or its signature.
To fix this situation, we implement the following changes:
- we no longer generate default implementation for missing constructor, only `declare` statement. These are ignored (like any other) during the Textual -> Sil translation.
- When a **function** call is made and there is a class of the same name (e.g. `B()` is a toplevel function call and we declared a class named `B`), we check if this class has a constructor. If it doesn't have one, we implement a default constructor in Pulse by allocating a value of the right type (`B` here) and then reuse method lookup resolution to locate the first `__init__` method in its hierarchy (`A.__init__` here).
- The way we encode `super().__init__()` will create static calls to `__init__` (e.g. `B.__init__(self)` in the example above for `[0]`) which is not defined. In this case, we also check if there's a class named `B` with an implicit constructor, and using method resolution to locate the right method to call.

Reviewed By: jvillard

Differential Revision: D48068959

fbshipit-source-id: 1133377ca62d561679cfbf59383c51238229682b