[infer/python] adding constructor return type information + fix method call type name
Summary:
In this diff we propagate the constructor return type information when we can. This will improve typing in Textual/Sil files.
We also fix a problem with method calls: since we don't know if a method is inherited or not (well, we might, but in general, we don't), we have to use the `?` name. Otherwise, Textual is looking exactly in the provided class, which is wrong, since the method might be inherited. Pulse / Tenv will do the resolution

Because we don't have type propagation yet, one test turns into a FN. The next diff will restore it.

Reviewed By: ngorogiannis

Differential Revision: D48189113

fbshipit-source-id: 5326211e1b5a49da86587c938e94364ce13ab575