[clang] Simplify the names of lambda classes
Summary:
The names given to the C++ lambdas by the compiler in the AST are way to long and convolute the readability of the SIL representation. It used to involve the method where the lambda is defined, and the (potentially) full path to the file with line and col.

Here we simplify it to be only "lambda_File.cpp:file:col", this should identify the lambdas while keeping the names short.

Reviewed By: skcho

Differential Revision: D48313606

fbshipit-source-id: cf01b7050750279967c4553fc0a25dd93c57c283