[clang] Create new names for the fields of lambdas that represent captured variables
Summary: We aim to unify the names for the lambda fields that appeared in the AST and are added to the tenv and used in copy constructors, with the names used by Pulse everywhere else. This diff adds new unified names in the frontend, and the following diff uses them in Pulse.

Reviewed By: skcho

Differential Revision: D48687675

fbshipit-source-id: 0e862658cd5fb67d336e657b9caea21f155b988c